Changed how sensor-target penalties are handled
Previously, it was handled on the target FIM level. One disadvantage, however, is that sometimes it would place a sensor close to a target and just register the contribution as 0 (as the math dictated). This obviously doesn't work, so now I look for the smallest target-sensor distance and apply that penalty globally.